+++ continuousintegration/orchestration/tests/changeStorageClass.sh
@@ -29,6 +29,7 @@ fi
 
 EOSINSTANCE=ctaeos
 NEW_STORAGE_CLASS_NAME=newStorageClassName
+TMP_DIR=$(mktemp -d)
 
 FILE_1=`uuidgen`
 FILE_2=`uuidgen`
@@ -36,16 +37,16 @@ echo
 echo "Creating files: ${FILE_1} ${FILE_2}"
 
 FRONTEND_IP=$(kubectl -n ${NAMESPACE} get pods ctafrontend -o json | jq .status.podIP | tr -d '"')
+kubectl -n ${NAMESPACE} exec ctafrontend -- bash -c "mkdir -p ${TMP_DIR}"
 
 echo
 echo "ENABLE CTAFRONTEND TO EXECUTE CTA ADMIN COMMANDS"
 kubectl --namespace=${NAMESPACE} exec kdc -- cat /root/ctaadmin2.keytab | kubectl --namespace=${NAMESPACE} exec -i ctafrontend --  bash -c "cat > /root/ctaadmin2.keytab; mkdir -p /tmp/ctaadmin2"
 kubectl -n ${NAMESPACE} cp client_helper.sh ctafrontend:/root/client_helper.sh
-rm /tmp/init_kerb.sh
-touch /tmp/init_kerb.sh
-echo '. /root/client_helper.sh; admin_kinit' >> /tmp/init_kerb.sh
-kubectl -n ${NAMESPACE} cp /tmp/init_kerb.sh ctafrontend:/tmp/init_kerb.sh
-kubectl -n ${NAMESPACE} exec ctafrontend -- bash /tmp/init_kerb.sh
+touch ${TMP_DIR}/init_kerb.sh
+echo '. /root/client_helper.sh; admin_kinit' >> ${TMP_DIR}/init_kerb.sh
+kubectl -n ${NAMESPACE} cp ${TMP_DIR}/init_kerb.sh ctafrontend:${TMP_DIR}/init_kerb.sh
+kubectl -n ${NAMESPACE} exec ctafrontend -- bash ${TMP_DIR}/init_kerb.sh
 
 
 echo
@@ -60,15 +61,16 @@ kubectl -n ${NAMESPACE} exec ctacli -- cta-admin sc ls
 
 echo
 echo "COPY REQUIRED FILES TO FRONTEND POD"
-echo "sudo kubectl cp ${NAMESPACE}/ctacli:/etc/cta/cta-cli.conf /etc/cta/cta-cli.conf"
-echo "sudo kubectl cp /etc/cta/cta-cli.conf ${NAMESPACE}/ctafrontend:/etc/cta/cta-cli.conf"
-sudo kubectl cp ${NAMESPACE}/ctacli:/etc/cta/cta-cli.conf /etc/cta/cta-cli.conf
-sudo kubectl cp /etc/cta/cta-cli.conf ${NAMESPACE}/ctafrontend:/etc/cta/cta-cli.conf
+echo "kubectl cp ${NAMESPACE}/ctacli:/etc/cta/cta-cli.conf ${TMP_DIR}/cta-cli.conf"
+echo "kubectl cp ${TMP_DIR}/cta-cli.conf ${NAMESPACE}/ctafrontend:/etc/cta/cta-cli.conf"
+kubectl cp ${NAMESPACE}/ctacli:/etc/cta/cta-cli.conf ${TMP_DIR}/cta-cli.conf
+kubectl cp ${TMP_DIR}/cta-cli.conf ${NAMESPACE}/ctafrontend:/etc/cta/cta-cli.conf
 
-kubectl -n ${NAMESPACE} cp common/archive_file.sh client:/usr/bin/
+kubectl -n ${NAMESPACE} exec client -- bash -c "mkdir -p ${TMP_DIR}"
+kubectl -n ${NAMESPACE} cp common/archive_file.sh client:${TMP_DIR}/
 kubectl -n ${NAMESPACE} cp client_helper.sh client:/root/
-kubectl -n ${NAMESPACE} exec client -- bash /usr/bin/archive_file.sh -f ${FILE_1} || exit 1
-kubectl -n ${NAMESPACE} exec client -- bash /usr/bin/archive_file.sh -f ${FILE_2} || exit 1
+kubectl -n ${NAMESPACE} exec client -- bash ${TMP_DIR}/archive_file.sh -f ${FILE_1} || exit 1
+kubectl -n ${NAMESPACE} exec client -- bash ${TMP_DIR}/archive_file.sh -f ${FILE_2} || exit 1
 
 EOS_METADATA_PATH_1=$(mktemp -d).json
 echo "SEND EOS METADATA TO JSON FILE: ${EOS_METADATA_PATH}"
@@ -85,17 +87,15 @@ EOS_ARCHIVE_ID_2=$(jq -r '.xattr | .["sys.archive.file_id"]' ${EOS_METADATA_PATH
 echo
 echo "CHANGE FILES WITH IDS"
 cd ~
-IDS_FILEPATH=~/archiveFileIds.txt
-rm ${IDS_FILEPATH}
+IDS_FILEPATH=${TMP_DIR}/archiveFileIds.txt
 touch ${IDS_FILEPATH}
 echo '{"archiveId": '${EOS_ARCHIVE_ID_1}'}' >> ${IDS_FILEPATH}
 echo '{"archiveId": '${EOS_ARCHIVE_ID_2}'}' >> ${IDS_FILEPATH}
 sleep 1
 
 echo
-kubectl cp ~/CTA-build/cmdline/standalone_cli_tools/change_storage_class/cta-change-storage-class ${NAMESPACE}/ctafrontend:/usr/bin/
-echo "kubectl cp ${IDS_FILEPATH} ${NAMESPACE}/ctafrontend:~/"
-kubectl cp ${IDS_FILEPATH} ${NAMESPACE}/ctafrontend:/root/
+echo "kubectl cp ${IDS_FILEPATH} ${NAMESPACE}/ctafrontend:${IDS_FILEPATH}"
+kubectl cp ${IDS_FILEPATH} ${NAMESPACE}/ctafrontend:${IDS_FILEPATH}
 echo "kubectl -n ${NAMESPACE} exec ctafrontend -- bash -c XrdSecPROTOCOL=krb5 KRB5CCNAME=/tmp/ctaadmin2/krb5cc_0 cta-change-storage-class --storageclassname ${NEW_STORAGE_CLASS_NAME} --json ${IDS_FILEPATH}"
 kubectl -n ${NAMESPACE} exec ctafrontend -- bash -c "XrdSecPROTOCOL=krb5 KRB5CCNAME=/tmp/ctaadmin2/krb5cc_0 cta-change-storage-class --storageclassname ${NEW_STORAGE_CLASS_NAME} --json ${IDS_FILEPATH} -t 1"